ref(recipe page - Recipe): move folders
