+++ src/app/recipes/[recipeKey]/Recipe/IngredientList/IngredientInformation/Product/style.module.scss
